requirements.txt
alpaca-py>=0.8.0
pandas>=1.5.0
numpy>=1.21.0
python-dotenv>=0.19.0 